% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_maturity.R
\name{compare_maturity}
\alias{compare_maturity}
\title{compare_maturity: R function to compute PIC, a probability for M1, M2 and M3 conditions by simulations}
\usage{
compare_maturity(n.iter = 1000, msurv0=10, msurv1=20, nevents=88, Ta=16.6, N=166, lfu=0, p1=0.5)
}
\arguments{
\item{n.iter:}{# of iterations, a positive integer value in (0, 100000)}

\item{msurv1:}{median survival time in treatment arm, a numeric value in (0, 100)}

\item{msurv0:}{median survival time in control arm, a numeric value in (0, 100)}

\item{nevents:}{# of events (can be computed by Ea() function),  a numeric value in (0, 1000)}

\item{Ta:}{accrual duration (can be computed by get_NTset() function), a numeric value in (0, 10000)}

\item{N:}{sample size (can be computed by get_NTset() function), a positive integer value in (nevents, 1000)}

\item{lfu:}{rate of loss-to-follow-up, a numeric value in (0, 1)}

\item{p1:}{allocation rate to treatment arm, a numeric value in (0, 1) (p0 = 1 - p1)}
}
\value{
T.m: mean of end of study time

Pic: a probability of an immature curve

M1: Proportion of events among all patients

M2: Median Follow-up among censored patiently only (m, sd)

M3: reverse KM for all patients (m, sd)
}
\description{
compare_maturity: R function to compute PIC, a probability for M1, M2 and M3 conditions by simulations
}
\references{
Joo L. et al. ``Data Maturity and its Implications in Confirmatory Trials with Time-to-Event Endpoints." (2018)
}
