% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E1.R
\name{E1}
\alias{E1}
\title{E1: R function to compute the number of events in one study arm under the PH assumption}
\usage{
E1(t = 24, Ta = 12, ra = 10, lambda =log(2)/10)
}
\arguments{
\item{t:}{time of analysis, a numeric value in (0, 1000)}

\item{Ta:}{accrual duration, a numeric value in (0, 1000)}

\item{ra:}{accural rate, a numeric value in (0, 100)}

\item{lambda:}{hazard rate, a numeric value in (0, 1)}
}
\value{
# of exepected events, an integer value (rounded-up)
}
\description{
E1: R function to compute the number of events in one study arm under the PH assumption
}
\references{
Kim and Tsiatis. ``Study duration for clinical trials with survival response and early stopping rule." Biometrics (1990): 81-92.
}
