% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E2.R
\name{E2}
\alias{E2}
\title{E2: R function to compute the number of events in two study arms under the PH assumption}
\usage{
E2(t = 24, Ta = 12, ra = 10, lambda1 = log(2)/20, lambda0 = log(2)/10)
}
\arguments{
\item{t:}{time of analysis, a numeric value in (0, 1000)}

\item{Ta:}{accrual time, a numeric value in (0, 1000)}

\item{ra:}{accural rate, a numeric value in (0, 100)}

\item{lambda1:}{hazard rate for treatment arm, a numeric value in (0, 1)}

\item{lambda0:}{hazard rate for control arm, a numeric value in (0, 1)}
}
\value{
# of exepected events, an integer value (rounding-up)
}
\description{
E2: R function to compute the number of events in two study arms under the PH assumption
}
\references{
Kim, Kyungmann, and Anastasios A. Tsiatis. ``Study duration for clinical trials with survival response and early stopping rule." Biometrics (1990): 81-92.
}
