% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ea.R
\name{Ea}
\alias{Ea}
\title{Ea: R function to compute the number of events to reject a log-rank test with given power and alpha.}
\usage{
Ea(HR=0.5, alpha=0.025, power=0.9, p1=0.5)
}
\arguments{
\item{HR:}{Hazard Ratio, a numeric value in (0, 1)}

\item{alpha:}{type I error of one-sided log-rank test, a numeric value in (0, 1)}

\item{power:}{power required to observe the significant result under current settings, a numeric value in (0, 1)}

\item{p1:}{allocation rate to treatment arm, a numeric value in (0, 1) (p0 = 1 - p1)}
}
\value{
# of exepected events, an integer value (rounding-up)
}
\description{
Ea: R function to compute the number of events to reject a log-rank test with given power and alpha.
}
\references{
Schoenfeld, David. "Sample-size formula for the proportional-hazards regression model." Biometrics (1983): 499-503.
}
