% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_N.R
\name{get_N}
\alias{get_N}
\title{get_N: R function to compute the number of events in two study arms under the PH assumption}
\usage{
get_N(Ts = 24, ra = 10, msurv1 = 20, msurv0 = 10)
}
\arguments{
\item{Ta:}{accrual duration,  a numeric value in (0, 1000)}

\item{ra:}{accural rate, a numeric value in (0, 100)}

\item{msurv1:}{median survival time in treatment arm, a numeric value in (0, 100)}

\item{msurv0:}{median survival time in control arm, a numeric value in (0, 100)}

\item{alpha:}{type I error of one-sided log-rank test, a numeric value in (0, 1)}

\item{power:}{power required to observe the significant result under current settings, a numeric value in (0, 1)}

\item{p1:}{allocation rate to treatment arm, a numeric value in (0, 1) (p0 = 1 - p1)}
}
\value{
N: sample size, a positive integer value (rounded-up)
}
\description{
get_N: R function to compute the number of events in two study arms under the PH assumption
}
\references{
Kim, Kyungmann, and Anastasios A. Tsiatis. ``Study duration for clinical trials with survival response and early stopping rule." Biometrics (1990): 81-92.
}
