% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTset.R
\name{get_NTset}
\alias{get_NTset}
\title{get_NTset: R function to generate a series of sets of N and T when ra, lambda0, lambda1 are given.}
\usage{
get_NTset(ra=10, msurv0=10, msurv1=20, alpha=0.025, power=0.9, p1=0.5, ta.d=0.1)
}
\arguments{
\item{ra:}{accural rate, a numeric value in (0, 100)}

\item{msurv1:}{median survival time in treatment arm, a numeric value in (0, 100)}

\item{msurv0:}{median survival time in control arm, a numeric value in (0, 100)}

\item{alpha:}{type I error of one-sided log-rank test, a numeric value in (0, 1)}

\item{power:}{power required to observe the significant result under current settings, a numeric value in (0, 1)}

\item{p1:}{allocation rate to treatment arm, a numeric value in (0, 1) (p0 = 1 - p1)}

\item{d:}{a number of decimal places (used for get_T), an positive integer value, default = 2}

\item{ta.d:}{Ta grid size , a numeric value in (0, 1)}
}
\value{
a tuple of (accrual duration (Ta), study duration (Ts), sample size (N))
}
\description{
get_NTset: R function to generate a series of sets of N and T when ra, lambda0, lambda1 are given.
}
\references{
Joo L. et al. ``Data Maturity and its Implications in Confirmatory Trials with Time-to-Event Endpoints." (2018)
}
