% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_T.R
\name{get_T}
\alias{get_T}
\title{get_T: R function to compute the study duration in two study arms under the PH assumption}
\usage{
get_T(Ta = 12, ra = 10, msurv1 = 20, msurv0 = 10, d=1)
}
\arguments{
\item{Ta:}{accrual duration, a numeric value in (0, 10000)}

\item{ra:}{accural rate, a numeric value in (0, 100)}

\item{msurv1:}{median survival time in treatment arm, a numeric value in (0, 100)}

\item{msurv0:}{median survival time in control arm, a numeric value in (0, 100)}

\item{alpha:}{type I error of one-sided log-rank test, a numeric value in (0, 1)}

\item{power:}{power required to observe the significant result under current settings, a numeric value in (0, 1)}

\item{p1:}{allocation rate to treatment arm, a numeric value in (0, 1) (p0 = 1 - p1)}

\item{d:}{a number of decimal places, an positive integer value, default = 1}
}
\value{
Ts, study duration, a numeric value in (0, 1000)
}
\description{
get_T: R function to compute the study duration in two study arms under the PH assumption
}
\references{
Kim, Kyungmann, and Anastasios A. Tsiatis. ``Study duration for clinical trials with survival response and early stopping rule." Biometrics (1990): 81-92.
}
