% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_sim.R
\name{naive_sim}
\alias{naive_sim}
\title{naive_sim: R function to simulate survival outcomes for a two arm study}
\usage{
naive.sim(nevents=88, lambda0=log(2)/10, lambda1=log(2)/20, Ta, N, lfu=0, p1=0.5)
}
\arguments{
\item{nevents:}{# of events (a value from Ea() function),  a numeric value in (0, 1000)}

\item{msurv1:}{median survival time in treatment arm, a numeric value in (0, 100)}

\item{msurv0:}{median survival time in control arm, a numeric value in (0, 100)}

\item{Ta:}{accrual duration (can compute with get_NTset() function), a numeric value in (0, 10000)}

\item{N:}{sample size (can compute with get_NTset() function), a positive integer value in (nevents, 1000)}

\item{lfu:}{rate of loss-to-follow-up, a numeric value in (0, 1)}

\item{p1:}{allocation rate to treatment arm, a numeric value in (0, 1) (p0 = 1 - p1)}
}
\value{
data= [group, time to enroll, time to observe, event.index (1=event, 0=censored)]
}
\description{
naive_sim: R function to simulate survival outcomes for a two arm study
}
\references{
Joo L. et al. ``Data Maturity and its Implications in Confirmatory Trials with Time-to-Event Endpoints." (2018)
}
