% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC_MC.R
\name{PIC_MC}
\alias{PIC_MC}
\title{PIC_MC: R function to compute PIC by MC}
\usage{
PIC_MC(n.iter = 10000, msurv0=10, msurv1=20, nevents=88, Ta=16.6, N=166, p1=0.5)
}
\arguments{
\item{n.iter:}{# of iterations, a positive integer value in (0, 10000)}

\item{msurv1:}{median survival time in treatment arm, a numeric value in (0, 100)}

\item{msurv0:}{median survival time in control arm, a numeric value in (0, 100)}

\item{nevents:}{# of events (can be computed by Ea() function),  a numeric value in (0, 1000)}

\item{Ta:}{accrual duration (can be computed by get_NTset() function), a numeric value in (0, 10000)}

\item{N:}{sample size (can be computed by get_NTset() function), a positive integer value in (nevents, 1000)}

\item{p1:}{allocation rate to treatment arm, a numeric value in (0, 1) (p0 = 1 - p1)}
}
\value{
a tuple of (mean of end of study time (T.m), sd of end of study time (T.sd), a probability of an immature curve (Pic))
}
\description{
PIC_MC: R function to compute PIC by MC
}
\references{
Joo L. et al. ``Data Maturity and its Implications in Confirmatory Trials with Time-to-Event Endpoints." (2018)
}
