% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare}
\alias{plot_compare}
\title{plot_compare: R function to draw the profiles of PIC and M1-M3 over designs}
\usage{
plot_compare(design, Pic=pic1, m1 = m1, m2 = m2, m3 = m3, n.plot.min=FALSE, n.plot.max=FALSE, t.plot.min=FALSE, t.plot.max=FALSE, col.d=2, col1=2, col2=3, col3=4, col.p="magenta")
}
\arguments{
\item{n.plot.min}{(optional) : adjusting the beginning of N grid (x-axis) in the plot}

\item{n.plot.max}{(optional) : adjusting the ending of N grid (x-axis) in the plot}

\item{t.plot.max}{(optional) : adjusting the ending of Ts grid (y-axis) in the plot}

\item{col.d}{: color used for design curve}

\item{col.p}{: color used for PIC curve}

\item{design:}{a tuple of three design parameters (Ta, Ts, N)}

\item{Pic:}{a vector of PIC under various desings of consideration}

\item{m1:}{a vector of M1 statistics under various desings of consideration}

\item{m2:}{a vector of M2 statistics under various desings of consideration}

\item{m3:}{a vector of M3 statistics under various desings of consideration}

\item{n.plot.min}{(optional) : adjusting the beginning of Ts grid (y-axis) in the plot}
}
\value{
a plot with
}
\description{
plot_compare: R function to draw the profiles of PIC and M1-M3 over designs
}
\references{
Joo L. et al. ``Data Maturity and its Implications in Confirmatory Trials with Time-to-Event Endpoints." (2018)
}
