% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pic.R
\name{plot_pic}
\alias{plot_pic}
\title{plot_pic: R function to draw the profile of PIC over designs}
\usage{
plot_pic(design, Pic=pic1, n.plot.min=FALSE, n.plot.max=FALSE, t.plot.min=FALSE, t.plot.max=FALSE,   col.d=2, col.p="magenta")
}
\arguments{
\item{n.plot.min}{(optional) : adjusting the beginning of N grid (x-axis) in the plot}

\item{n.plot.max}{(optional) : adjusting the ending of N grid (x-axis) in the plot}

\item{t.plot.max}{(optional) : adjusting the ending of Ts grid (y-axis) in the plot}

\item{col.d}{: color used for design curve}

\item{col.p}{: color used for PIC curve}

\item{design:}{a tuple of three design parameters (Ta, Ts, N)}

\item{Pic:}{a vector of PIC under various desings of consideration}

\item{n.plot.min}{(optional) : adjusting the beginning of Ts grid (y-axis) in the plot}
}
\value{
a plot of two curves: one for (N, Ts) and the other for (N, PIC)
}
\description{
plot_pic: R function to draw the profile of PIC over designs
}
\references{
Joo L. et al. ``Data Maturity and its Implications in Confirmatory Trials with Time-to-Event Endpoints." (2018)
}
